refactor: extract Pagination as a reusable component
